Work on Favorites Seciotn. Ref #12. Ref #13.
Implemented some functions to assist in the building of the favorites section (and the file libraries). Work on function called favorites that will do more with the database. For now it checks to see if a file is "favorited" by the session student and if it IS then the Favorite! button changes to Favorited.