Add explanation to Favorite's button logic. Close #13
Add some explanation to the more (recently) complicated functions and bits of code that may be hard to understand associated with the Favorites button. It may not be the most effecient way of doing what I want (I know it's not) but to make clicking a button do so much wasn't as easy as I thought it would be.