Take out a function that makes some HTML. Ref #16
The created Student object in the navbar doesn't carry over to the functions file and it complained if i created a newer new Student object within the functions file. So i took out an HTML function that just echos out some html tags. Fixed the files that used this function by just adding the simple HTML back in with the new Class getter