Moving student object to different file. Close #24. Ref #16
Also moving Database object to functions file. Idea is to get rid of all the global functions in the function.php file and instead just include the class files that will house the deleted functions as methods. This will be a common file included in every file so all files will have access to these classes and objects easily.