Work on Class System for Content libraries. Ref #25
Created the Content Class; used to create Content objects such as: Etudes, Audio Files, and Videos. Each object will have access to methods that get specific information from the Database about them. For this, an extension of the Database Class was made so I could access the query methods inside the DB Class. 
Also, tweaking was needed in the Database Class to call a private test function.