Rework/Vastly Improve Database Class. Close #28. Close #30
The previous way of implementing the DB Class had one constructor that just set a property to the new connection. This new way introduces an $instance property that insures only ONE DB connection can be made (See the connect() method) From here, there is a setServer method that can be statically called to give the DB Class the correct DB creds. From here, I have created more query methods and fixed an issue with trying to statically call a method instead of using the $this DB object to do so.