Implement Central file with Class/Object access. Close #31
This file is now used to house all the file paths to the Class files as well as being the location where the Student Object "lives." This file is included in almost all files so it A) eliminates the need for multiple files to have multiple include statements, and B) since it is included in most files, I don't have to recreate a Student object everytime I move pages. This is much more cleaner and more organized way and I prefer this implementation of a "functions" file; rather than have hundreds of lines of functions.