Fix login system to use the Class system. Ref #32
Instead of having the file do all the work with the queries, I offload those processes to the DB Class that handles all query work. I create a temporary Student object (with the information the user entered in the login forum) as a checker against the information I have in the Database. If the two match, the user is logged in; setting the SESSION variables so the main functions file can build the real Student Object.