Fix update user process to use th Class System. Ref #32
Same way with the login process; use the DB Class to do the actual update query work. Only difference is this file used the Student Object in the functions file. If an update was successful, then it uses the student class's setter methods to update the Student Object instead of creating a whole new student object. The html file doesn't need any include statements since it is already being used by the php file, so all I have to do is update the placeholder spots to use the Student Object's getter methods.