Moved include path for the server details. Ref #16
There was no need to have the file path to the dummy_server/server details array in the functions file, so i moved it directly to the DB Class file. This in turn broke the Login process so I had to grab the DB instance method from the DB class to perform the login process correctly.