Work on displaying a Student's Favorites. Close #12. Ref #17
Created a new file called 'display_favorites' which has very similar functionality to the 'library_builder' file only with some important changes.  I couldn't create a Content Object as easily so I have to build one based on the Info I get back from my DB query that is in an Array data structure,  I also found it much simpler to have a separate file do all the display work and then just include that into the index file. I had to move the include down passed the <head> tag in order for the content to be displayed correctly, otherwise the Navbar would cut off a lot of the first row.