Work on back-end Implementation for Favorite Button. Ref #34
Added a new query method to the DB Class. It is given a file and checks to see if that file exists in the Favorites table under the given student ID. If it is not there, then it is not Favorited. Updated student class to call this method. Updated the old add_Favorites file to handle when the Favorite button is pressed and what info to get and add to the database.