Correctly display the Favorite Button. Close #35. Close #34
Updated the Twig templates to include information about if they are Favorited, based on the new checkStudentFavorite method in the DB class. "Submitting" the Favorite button form on an unfavorited file loads the add_favorites file, adds the new obj to the favorites table, then reloads the page you were on; updating the Favroite button and adding it to the index homepage under the student's favorite secrion.