Work on DB Methods for Student's Rudiment Progress. Ref #33
Implement a set method that initializes 40 rows in the Student Progress table associated with the current Student. 
Another methd to check a single row of that Progress table, more specificlly the 'status' field. '0' means it is "uncompleted" and a '1' means "completed."
Lastly, an update method to change the 'status' field of one row from '0' to '1,' marking that rudiment "completed."