Cleanup and adjustments for Favorites option/section. Close #40
deleted the 'add_Favorites' file for a universal 'favorites' file that handles both Favoriting and UNfavoriting of files. Also, add in more variables to the Twig template loader on the display_favorites file to match the Library_builder file.

This no results in a working Unfavorite option. Clicking on the Favorited! button of a already Favorited file will UNfavorite that file.