Fix implementation of Student Progress. Ref #33
Originally I had intended to "setup" a student's progress table to have 40 rows (representing the 40 rudiments) and set one of their fields to 0, meaning uncomplete. As they go through a click the html button to complete a rudiment, it would go in and just change one field in one row. I realize having 40 rows in the beginning is overcrowding the DB so I set it up to check to see IF there is a row in the table, if there is not: then that rudiment is "uncomplete" and if there IS a row (matching the parameters) then it will be "complete."

Also, Commit 100!